transport: convert pre-push hook to hook.h
Move the pre-push hook away from run-command.h to and over to
the new hook.h library.

In addition to moving the signal and process handling, we use
a custom stdin feeding callback which preserves the original
behaviour of feeding stdin line by line, to avoid buffering
too much data in memory.

Signed-off-by: Emily Shaffer <emilyshaffer@google.com>
Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>