submodule: add gitdir path config override
This adds the ability to override gitdir paths via config files
(not .gitmodules) such that the encoding scheme (or plain text
name if the encoding extension is disabled) can be changed via
config entries.

These entries are not added by default for all submodules: they
should be used on an as-needed basis.

A new test and a helper are added. The helper will also be used
in further tests exercising gitdir encoding functionality.

Based-on-patch-by: Brandon Williams <bmwill@google.com>
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>