submodule: error out if gitdir name is too long
Encoding submodule names increases their name size, so there is an
increased risk to hit the max filename length in the gitdir path.

This gitdir file-name-too-long corner case can be be addressed in
multiple ways (including sharding or trimming), however, for now,
just add portable logic to detect and error out, to avoid comitting
to a specific policy and keeping backwards compatibility for it.

Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>