hook: allow overriding ungroup option
When calling run_process_parallel() in run_hooks_opt(), the
ungroup option is currently hardcoded to .ungroup = 1.

This causes problems when ungrouping should be disabled, for
example when sideband-reading collated output from child hooks,
because sideband-reading and ungrouping are mutually exclusive.

Thus we add a new hook.h option which allows overriding it.

Existing ungroup=1 behavior is kept as is in the run_hooks()
API and the "hook run" command. We could modify these to take
an option if necessary, so I left two code comemnts there.

Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>