hook: allow overriding the ungroup option
When calling run_process_parallel() in run_hooks_opt(), the
ungroup option is currently hardcoded to .ungroup = 1.

This causes problems when ungrouping should be disabled, for
example when sideband-reading collated output from child hooks,
because sideband-reading and ungrouping are mutually exclusive.

Thus a new hook.h option is added to allow overriding.

The existing ungroup=1 behavior is preserved as the default via
the struct run_hooks_opt initializer and the default is changed
only in selective cases where needed (see next commits which
convert receive-pack).

Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>