submodule: use hashed name for gitdir
If none of the previous steps work and we reach case 2.4, try to
hash the submodule name and see if that can be a valid gitdir
before giving up and throwing an error.

This is a "last resort" type of measure to avoid conflicts since
it loses the gitdir human readability. Itis not such a big deal
because users are now supposed to use the submodule.<name>.gitdir
config as the single source of truth for gitdir paths.

This logic will be reached in very rare cases, as can be seen in
the test we added.

Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>