ws: add new tab-between-non-ws check
This adds a new check to detect HT in the middle of sentences that
should have been a SP, as suggested by Junio in
https://public-inbox.org/git/xmqqy0mwsedz.fsf@gitster.g/

The check is a bit complex because we want to detect places where
a SP was intended and the naive before/after character check can
issue false positives in cases like "a\tb".

The new check is enabled for Documentation/**/*.adoc, where these
kinds of mistakes were seen in practice.

Suggested-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>